rich text editor changes - cleanup old leftover codes - remove unecessary service dependency
