+++ src/Squidex/app/shared/components/asset.component.ts
@@ -45,10 +45,7 @@ export class AssetComponent implements OnInit {
 
     @Input()
     public closeMode = false;
-
-    @Input()
-    public hideIcons = false;
-
+    
     @Output()
     public loaded = new EventEmitter<AssetDto>();
